Common ServiceState api
