+++ protos/farm_ng/service/service.proto
@@ -1,0 +1,47 @@
+// Copyright (c) farm-ng, inc.
+//
+// Licensed under the Amiga Development Kit License (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://github.com/farm-ng/amiga-dev-kit/blob/main/LICENSE
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+syntax = "proto3";
+
+package farm_ng.service.proto;
+
+// TODO: Just use a bool where this is used?
+enum ReplyStatus {
+  OK = 0;
+  FAILED = 1;
+}
+
+enum ServiceState {
+  // Service state is not known
+  UNKNOWN = 0;
+  // Stopped
+  STOPPED = 1;
+  // Running, and actively serving data
+  RUNNING = 2;
+  // Idle, waiting for clients
+  IDLE = 3;
+  // Service isn't running at all
+  UNAVAILABLE = 4;
+  // Service is in some error state
+  ERROR = 5;
+}
+
+message GetServiceStateRequest {
+}
+
+message GetServiceStateReply {
+  ServiceState state = 1;
+  uint32 pid = 2; // process id
+  double uptime = 3;
+}
