+++ setup.cfg
@@ -43,6 +43,7 @@ packages =
     farm_ng
     farm_ng.canbus
     farm_ng.oak
+    farm_ng.service
 
 [options.extras_require]
 dev =
@@ -63,14 +64,11 @@ files = py/sophus, py/tests, py/examples
 ignore_missing_imports = True
 
 [options.package_data]
-farm_ng.controller =
-    *.proto
-
 farm_ng.oak =
     *.proto
 
 farm_ng.canbus =
     *.proto
 
-farm_ng.state_estimator =
+farm_ng.service =
     *.proto
